Move dtmpl_file and dflt_file tests
Move these tests from 'test_plugin_hydrotrend.py' to
'test_plugin_base.py'.
